#112 : Moved Email Alert Message builder to mail package
